LayoutTests/platform/wpe/accessibility/th-as-title-ui-expected.txt

CONSOLE MESSAGE: TypeError: null is not an object (evaluating 'group.titleUIElement().allAttributes')
Header for button	
Header for button	
